+++ oConsole.py
@@ -1,10 +1,35 @@
 import codecs, ctypes, ctypes.wintypes, locale, sys, threading;
 #oStdOut = codecs.getwriter(locale.getpreferredencoding())(sys.stdout, "replace");
 
+SHORT = ctypes.c_short;
+WORD = ctypes.c_ushort;
 STD_INPUT_HANDLE = -10
 STD_OUTPUT_HANDLE = -11
 STD_ERROR_HANDLE = -12
 NORMAL_COLOR = 0x7;
+
+class COORD(ctypes.Structure):
+  _fields_ = [
+    ("X", SHORT),
+    ("Y", SHORT)
+  ];
+
+class SMALL_RECT(ctypes.Structure):
+  _fields_ = [
+    ("Left", SHORT),
+    ("Top", SHORT),
+    ("Right", SHORT),
+    ("Bottom", SHORT)
+  ];
+    
+class CONSOLE_SCREEN_BUFFER_INFO(ctypes.Structure):
+  _fields_ = [
+    ("dwSize", COORD),
+    ("dwCursorPosition", COORD),
+    ("wAttributes", WORD),
+    ("srWindow", SMALL_RECT),
+    ("dwMaximumWindowSize", COORD)
+  ];
 
 # Non-unicode strings are assumed to be CP437. We have an indexed table to
 # convert CP437 to unicode (index range 0-255 => unicode char) and a dict to
@@ -63,12 +88,30 @@ class cConsole(object):
     dwMode = ctypes.wintypes.DWORD(0);
     oConsole.bStdOutIsConsole = ctypes.windll.kernel32.GetConsoleMode(oConsole.hStdOut, ctypes.byref(dwMode));
     oConsole.bByteOrderMarkWritten = False;
-
+    oConsole.uDefaultColor = -1;
+  
+  @property
+  def uCurrentColor(oConsole):
+    return oConsole.__fuGetCurrentColor();
+  
+  def __fuGetCurrentColor(oConsole):
+    oConsoleScreenBufferInfo = CONSOLE_SCREEN_BUFFER_INFO()
+    assert ctypes.windll.kernel32.GetConsoleScreenBufferInfo(oConsole.hStdOut, ctypes.byref(oConsoleScreenBufferInfo)), \
+        "GetConsoleScreenBufferInfo(%d, ...) => Error %08X" % \
+        (oConsole.hStdOut, ctypes.windll.kernel32.GetLastError());
+    return oConsoleScreenBufferInfo.wAttributes & 0xFF;
+  
   def __fSetColor(oConsole, uColor):
     assert oConsole.bStdOutIsConsole, \
         "Cannot set colors when output is redirected";
-    assert uColor in xrange(0, 0x100), \
+
+    uMask = uColor >> 8;
+    assert uMask in xrange(0, 0x100), \
         "You cannot use color 0x%X; maybe you are trying to print a number without converting it to a string?" % uColor;
+    uColor &= 0xFF;
+    if uMask:
+      uCurrentColor = oConsole.uCurrentColor;
+      uColor = (uCurrentColor & (uMask ^ 0xFF)) + (uColor & uMask);
     assert ctypes.windll.kernel32.SetConsoleTextAttribute(oConsole.hStdOut, uColor), \
         "SetConsoleTextAttribute(%d, %d) => Error %08X" % \
         (oConsole.hStdOut, uColor, ctypes.windll.kernel32.GetLastError());
@@ -95,16 +138,20 @@ class cConsole(object):
     global goLock, guLastLineLength;
     oConsole.oLock.acquire();
     try:
+      uOriginalColor = oConsole.uCurrentColor;
       # Go to the start of the current line if needed
       if oConsole.uLastLineLength:
         oConsole.__fWriteOutput(oConsole.bStdOutIsConsole and u"\r" or "\r");
       bColorWasSet = False;
+      if oConsole.uDefaultColor != -1:
+        oConsole.__fSetColor(oConsole.uDefaultColor);
+        bColorWasSet = True;
       uNewLineLength = 0;
       try:
         for xCharsOrColor in axCharsAndColors:
           if isinstance(xCharsOrColor, int) or isinstance(xCharsOrColor, long):
             if oConsole.bStdOutIsConsole: # If output is redirected, colors will not be set, so don't try
-              if xCharsOrColor == -1: xCharsOrColor = NORMAL_COLOR;
+              if xCharsOrColor == -1: xCharsOrColor = uOriginalColor;
               oConsole.__fSetColor(xCharsOrColor);
               bColorWasSet = True;
           else:
@@ -114,7 +161,7 @@ class cConsole(object):
             oConsole.__fWriteOutput(xCharsOrColor);
       finally:
         if bColorWasSet:
-          oConsole.__fSetColor(NORMAL_COLOR);
+          oConsole.__fSetColor(uOriginalColor);
       if uNewLineLength < oConsole.uLastLineLength:
         oConsole.__fWriteOutput(u" " * (oConsole.uLastLineLength - uNewLineLength));
       if bNewLine: